debian-pkg: Add `dist` as conflict
The package `dist` is a set of tools for developing, maintaining and
distributing software. Unfortunately, one of the scripts in this package
is named pat and installs to /usr/bin/pat (which is in conflict with our
package).

Defining `dist` as a conflicting package ensures that dpkg will refuse
to to allow them to be unpacked on the system at the same time.